[android] Attempts to fix crash from deleted bookmark
App crashes when removing bookmark and returning to map after pressing
POI and bottom menu bookmarks at the same time. The crash happens
because after the delete, the app tries to find the deleted bookmark id,
which does not exist. The best solution I found, is to make so if the app
wants to use the getBookmark function and it doesnt find it, to get the
last bookmark present in the container, this eliminates the bug but only if
there are more than one bookmarks present on the container.

fixes #6217

Signed-off-by: Francisco Nael Salgado <francisco.nael.salgado@tecnico.ulisboa.pt>